+++ src/javascript/app/base/binary_pages.js
@@ -16,6 +16,7 @@ const MBTradePage             = require('../pages/mb_trade/mb_tradepage');
 const EconomicCalendar        = require('../pages/resources/economic_calendar/economic_calendar');
 const AssetIndexUI            = require('../pages/resources/asset_index/asset_index.ui');
 const TradingTimesUI          = require('../pages/resources/trading_times/trading_times.ui');
+const Signup                  = require('../pages/signup');
 const TradePage               = require('../pages/trade/tradepage');
 const Authenticate            = require('../pages/user/account/authenticate');
 const ChangePassword          = require('../pages/user/account/change_password');
@@ -41,7 +42,6 @@ const MetaTrader              = require('../pages/user/metatrader/metatrader');
 const FinancialAccOpening     = require('../pages/user/new_account/financial_acc_opening');
 const JapanAccOpening         = require('../pages/user/new_account/japan_acc_opening');
 const RealAccOpening          = require('../pages/user/new_account/real_acc_opening');
-const Signup                  = require('../../static/pages/signup');
 const VirtualAccOpening       = require('../pages/user/new_account/virtual_acc_opening');
 const WelcomePage             = require('../pages/user/new_account/welcome_page');
 const ResetPassword           = require('../pages/user/reset_password');
